backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4

# NLP
spacy==3.7.5
nltk==3.8.1
scikit-learn==1.5.1
sentence-transformers==3.0.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.2.2
httpx==0.27.0
